luci.mk: use logical operators for install recipe
The use of bash if statements has extra syntax requirements
compared to logical operators && and ||
for example, that each line must end with a semicolon.

The default definition of Build/Install/Default
is not managed within the luci repo,
so a syntax requirement is not ideal.
A missing or extra semicolon for whatever reason
can cause build failure.

Also, this makes the copying from PKG_INSTALL_DIR
dependent on Build/Install/Default being successful,
where a bash if statement would still execute the copy otherwise,
for example, if the default "make install" fails.

Signed-off-by: Michael Pratt <mcpratt@pm.me>